update token with test_app

+++ test_app.py
@@ -16,6 +16,10 @@ class TriviaTestCase(unittest.TestCase):
         self.database_path = os.environ['DATABASE_URL'] = "postgresql://localhost:5432/zacademy_test"
         setup_db(self.app, self.database_path)
 
+        # TOKEN 
+        self.instructor_token = os.environ['instructor_token']
+        self.admin_token = os.environ['admin_token']
+        
         # binds the app to the current context
         with self.app.app_context():
             self.db = SQLAlchemy()
@@ -54,6 +58,19 @@ class TriviaTestCase(unittest.TestCase):
         """Executed after reach test"""
         pass
 
+# Auth test
+    def test_403_instructor(self):
+        res = self.client().post('/instructor', json=self.new_instructor, headers={"Authorization": 'bearer ' + self.instructor_token})
+        data = json.loads(res.data)
+        self.assertEqual(res.status_code, 403)
+        self.assertEqual(data['success'], False)
+        
+    def test_401_instructor(self):
+        res = self.client().post('/instructor', json=self.new_instructor)
+        data = json.loads(res.data)
+        self.assertEqual(res.status_code, 401)
+        self.assertEqual(data['success'], False)
+        
     def test_get_instructors(self):
         res = self.client().get('/instructors')
         data = json.loads(res.data)
@@ -61,31 +78,31 @@ class TriviaTestCase(unittest.TestCase):
         self.assertEqual(data['success'], True)
 
     def test_add_instructor(self):
-        res = self.client().post('/instructor', json=self.new_instructor)
+        res = self.client().post('/instructor', json=self.new_instructor, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_500_server_error_instructor(self):
-        res = self.client().post('/instructor', json=self.new_instructor_bad_req)
+        res = self.client().post('/instructor', json=self.new_instructor_bad_req, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 500)
         self.assertEqual(data['success'], False)
 
     def test_edit_instructor(self):
-        res = self.client().patch('/instructor/1', json=self.edit)
+        res = self.client().patch('/instructor/1', json=self.edit, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_404_if_instructor_not_exist(self):
-        res = self.client().delete('/instructor/1000', json=self.edit)
+        res = self.client().delete('/instructor/1000', json=self.edit, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
 
     def test_404_remove_instructor_not_exist(self):
-        res = self.client().delete('/instructor/1000')
+        res = self.client().delete('/instructor/1000', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
@@ -98,31 +115,31 @@ class TriviaTestCase(unittest.TestCase):
         self.assertEqual(data['success'], True)
 
     def test_add_course(self):
-        res = self.client().post('/course', json=self.new_course)
+        res = self.client().post('/course', json=self.new_course, headers={"Authorization": 'bearer ' + self.instructor_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_500_server_error_course(self):
-        res = self.client().post('/course', json=self.new_instructor_bad_req)
+        res = self.client().post('/course', json=self.new_instructor_bad_req, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 500)
         self.assertEqual(data['success'], False)
 
     def test_edit_course(self):
-        res = self.client().patch('/course/1', json=self.edit_type)
+        res = self.client().patch('/course/1', json=self.edit_type, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_404_if_course_not_exist(self):
-        res = self.client().delete('/course/1000', json=self.edit_type)
+        res = self.client().delete('/course/1000', json=self.edit_type, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
 
     def test_404_remove_course_not_exist(self):
-        res = self.client().delete('/course/1000')
+        res = self.client().delete('/course/1000', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
@@ -135,26 +152,26 @@ class TriviaTestCase(unittest.TestCase):
         self.assertEqual(data['success'], True)
 
     def test_create_class(self):
-        res = self.client().post('/class', json=self.new_class)
+        res = self.client().post('/class', json=self.new_class, headers={"Authorization": 'bearer ' + self.instructor_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_500_server_error_class(self):
-        res = self.client().post('/class', json=self.new_instructor_bad_req)
+        res = self.client().post('/class', json=self.new_instructor_bad_req, headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 500)
         self.assertEqual(data['success'], False)
 
     def test_404_remove_course_not_exist(self):
-        res = self.client().delete('/course/1000')
+        res = self.client().delete('/course/1000', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
     # Remove
 
     def test_remove_class(self):
-        res = self.client().delete('/class/1')
+        res = self.client().delete('/class/1', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         cou = Course.query.get(1)
         self.assertEqual(res.status_code, 200)
@@ -163,7 +180,7 @@ class TriviaTestCase(unittest.TestCase):
         self.assertEqual(cou, None)
 
     def test_remove_course(self):
-        res = self.client().delete('/course/1')
+        res = self.client().delete('/course/1', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         cou = Course.query.get(1)
         self.assertEqual(res.status_code, 200)
@@ -172,7 +189,7 @@ class TriviaTestCase(unittest.TestCase):
         self.assertEqual(cou, None)
 
     def test_remove_instructor(self):
-        res = self.client().delete('/instructor/1')
+        res = self.client().delete('/instructor/1', headers={"Authorization": 'bearer ' + self.admin_token})
         data = json.loads(res.data)
         inst = Instructor.query.get(1)
         self.assertEqual(res.status_code, 200)
